# frontend/shared/api/openapi.json
{
  "components": {
    "schemas": {
      "Body_delete_document_endpoint_api_documents_delete_post": {
        "properties": {
          "csrf_token": {
            "default": "",
            "title": "Csrf Token",
            "type": "string"
          },
          "relative_raw_path": {
            "default": "",
            "title": "Relative Raw Path",
            "type": "string"
          }
        },
        "title": "Body_delete_document_endpoint_api_documents_delete_post",
        "type": "object"
      },
...
      "Body_upload_document_api_documents_upload_post": {
        "properties": {
          "csrf_token": {
            "default": "",
            "title": "Csrf Token",
            "type": "string"
          },
          "file": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "File",
            "type": "array"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_documents_upload_post",
        "type": "object"
      },
...
      "ChunkDto": {
        "properties": {
          "char_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Char End"
          },
          "char_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Char Start"
          },
          "characters": {
            "title": "Characters",
            "type": "integer"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "index": {
            "title": "Index",
            "type": "integer"
          }
        },
        "required": [
          "index",
          "char_start",
          "char_end",
          "characters",
          "content"
        ],
        "title": "ChunkDto",
        "type": "object"
      },
...
      "DocumentChunksResponse": {
        "properties": {
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/ChunkDto"
            },
            "title": "Chunks",
            "type": "array"
          },
          "document": {
            "$ref": "#/components/schemas/DocumentDto"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "document",
          "chunks",
          "page",
          "page_size",
          "total"
        ],
        "title": "DocumentChunksResponse",
        "type": "object"
      },
      "DocumentDeleteRequest": {
        "properties": {
          "document_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "minItems": 1,
            "title": "Document Ids",
            "type": "array"
          }
        },
        "required": [
          "document_ids"
        ],
        "title": "DocumentDeleteRequest",
        "type": "object"
      },
      "DocumentDto": {
        "properties": {
          "byte_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Byte Size"
          },
          "content_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Hash"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "indexed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Indexed"
          },
          "indexed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Indexed At"
          },
          "original_filename": {
            "title": "Original Filename",
            "type": "string"
          },
          "parsed_exists": {
            "title": "Parsed Exists",
            "type": "boolean"
          },
          "parsed_markdown_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parsed Markdown Path"
          },
          "processed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed At"
          },
          "processing_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Error"
          },
          "raw_exists": {
            "title": "Raw Exists",
            "type": "boolean"
          },
          "relative_raw_path": {
            "title": "Relative Raw Path",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "relative_raw_path",
          "original_filename",
          "display_name",
          "byte_size",
          "raw_exists",
          "parsed_markdown_path",
          "parsed_exists",
          "content_hash",
          "processed_at",
          "indexed",
          "status"
        ],
        "title": "DocumentDto",
        "type": "object"
      },
      "DocumentIndexResponse": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentDto"
          },
          "index": {
            "$ref": "#/components/schemas/IndexPreviewDto"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "document",
          "index"
        ],
        "title": "DocumentIndexResponse",
        "type": "object"
      },
      "DocumentMarkdownResponse": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentDto"
          },
          "markdown": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Markdown"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "document",
          "markdown"
        ],
        "title": "DocumentMarkdownResponse",
        "type": "object"
      },
      "DocumentPageResponse": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentDto"
            },
            "title": "Documents",
            "type": "array"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "qdrant_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qdrant Error"
          },
          "status_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Status Counts",
            "type": "object"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "documents",
          "page",
          "page_size",
          "total",
          "pages",
          "status_counts",
          "qdrant_error"
        ],
        "title": "DocumentPageResponse",
        "type": "object"
      },
      "DocumentResponse": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/DocumentDto"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "document"
        ],
        "title": "DocumentResponse",
        "type": "object"
      },
      "DocumentUploadResponse": {
        "properties": {
          "failed_count": {
            "title": "Failed Count",
            "type": "integer"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/UploadResultDto"
            },
            "title": "Results",
            "type": "array"
          },
          "uploaded_count": {
            "title": "Uploaded Count",
            "type": "integer"
          },
          "uploads": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Uploads",
            "type": "array"
          }
        },
        "required": [
          "ok",
          "uploads",
          "results",
          "uploaded_count",
          "failed_count"
        ],
        "title": "DocumentUploadResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UploadResultDto": {
        "properties": {
          "bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bytes"
          },
          "document_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "file_name": {
            "title": "File Name",
            "type": "string"
          },
          "relative_raw_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relative Raw Path"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "file_name",
          "status"
        ],
        "title": "UploadResultDto",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Medic RAG Dashboard",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/documents": {
      "get": {
        "operationId": "api_documents_api_documents_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "updated_at",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "default": "desc",
              "title": "Direction",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Documents"
      }
    },
    "/api/documents/delete": {
      "post": {
        "operationId": "delete_document_endpoint_api_documents_delete_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_delete_document_endpoint_api_documents_delete_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Document Endpoint"
      }
    },
    "/api/documents/delete-by-id": {
      "post": {
        "operationId": "delete_documents_by_id_endpoint_api_documents_delete_by_id_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Documents By Id Endpoint"
      }
    },
...
    "/api/documents/process": {
      "get": {
        "operationId": "api_document_process_api_documents_process_get",
        "parameters": [
          {
            "in": "query",
            "name": "relative_raw_path",
            "required": true,
            "schema": {
              "title": "Relative Raw Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Api Document Process Api Documents Process Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Document Process"
      }
    },
    "/api/documents/upload": {
      "post": {
        "operationId": "upload_document_api_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Document"
      }
    },
    "/api/documents/{document_id}": {
      "get": {
        "operationId": "api_document_api_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Document"
      }
    },
    "/api/documents/{document_id}/chunks": {
      "get": {
        "operationId": "api_document_chunks_api_documents__document_id__chunks_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentChunksResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Document Chunks"
      }
    },
    "/api/documents/{document_id}/index-points": {
      "get": {
        "operationId": "api_document_index_points_api_documents__document_id__index_points_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentIndexResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Document Index Points"
      }
    },
    "/api/documents/{document_id}/markdown": {
      "get": {
        "operationId": "api_document_markdown_api_documents__document_id__markdown_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentMarkdownResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Document Markdown"
      }
    },
...
  }
}